// packages/sdk-cloudsave/src/generated-public/cloudsave/definitions/PlayerRecordResponse.ts
/*
 * Copyright (c) 2022-2023 AccelByte Inc. All Rights Reserved
 * This is licensed software from AccelByte Inc, for limitations
 * and restrictions contact your company contract manager.
 */
import { z } from 'zod'

export const PlayerRecordResponse = z.object({
  created_at: z.string(),
  is_public: z.boolean(),
  key: z.string(),
  namespace: z.string(),
  set_by: z.string().nullish(),
  updated_at: z.string(),
  user_id: z.string(),
  value: z.record(z.any())
})

export interface PlayerRecordResponse extends z.TypeOf<typeof PlayerRecordResponse> {}